refactor(web): simplify gate and toast shortcut wiring
Move the digit-1 binding into ToastActionButton and share letter-key dispatch plus gate panel styles between trial and billing overlays.
